builder/requirements.txt
ruamel.yaml
jina==0.1.8
